backend/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
idna==3.7
pillow==10.2.0
PyJWT==1.7.1
pytz==2023.3.post1
requests==2.32.2
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.2.1
